% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ci}
\alias{get_ci}
\title{Get the confidence interval for a given effect size}
\usage{
get_ci(es, level = 0.95, df = NA)
}
\arguments{
\item{es}{A data frame of the class mpp with an effect size and its variance.
For example, the output of \code{\link{md}}, \code{\link{smd}},
\code{\link{scr}}.}

\item{level}{the confidence level. Defaults to 0.95.}

\item{df}{The degrees of freedom. Defaults to NA, which assumes a normal
distribution. Otherwise a t-distribution is assumed.}
}
\value{
A data.frame with the effect size (es), its variance (var) and the
lower and upper limits of the confidence interval (ci_low, ci_high).
}
\description{
Calculates the confidence interval for
}
\examples{
# calculate a SMD first
g <- smd(x1 = 10, x2 = 15, sd1 = 3, sd2 = 4, n1 = 11, n2 = 9)

# calculate confidence interval
get_ci(g)

}
