% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{j}
\alias{j}
\title{Calculating the Hedges correction factor}
\usage{
j(x, exact = TRUE)
}
\arguments{
\item{x}{numeric, the degrees of freedom.}

\item{exact}{Whether to use the exact formula or an approximation. Defaults
to true.}
}
\value{
A numeric, the bias correction factor.
}
\description{
Calculates the Hedges correction factor for small sample bias
}
\examples{
j(20)

}
