% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{md}
\alias{md}
\title{Calculating a mean difference}
\usage{
md(x1, x2, sd1, sd2, n1, n2, var_homo = TRUE)
}
\arguments{
\item{x1, x2}{means of the first and seconds group.}

\item{sd1, sd2}{standard deviations of the first and second group.}

\item{n1, n2}{sample sizes of the first and second group.}

\item{var_homo}{whether an variance estimator with pooled standard deviation
should be used, that assumes homoscedasticity for the population variance
(defaults to TRUE).}
}
\value{
A data frame of the class mpp with the effect size (es) and its
variance (var).
}
\description{
Calculates a raw MD effect size for a simple group comparison.
}
\examples{
md(x1 = 10, x2 = 15, sd1 = 3, sd2 = 4, n1 = 11, n2 = 9)

}
