% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pool_groups}
\alias{pool_groups}
\title{Pool groups}
\usage{
pool_groups(x1, x2, sd1, sd2, n1, n2)
}
\arguments{
\item{x1, x2}{means of the first and seconds group.}

\item{sd1, sd2}{standard deviations of the first and second group.}

\item{n1, n2}{sample sizes of the first and second group.}
}
\value{
a data frame with the group mean (mean), standard deviation (sd) and sample size (n).
}
\description{
Pools the mean, standard deviation and sample size for two groups.
}
\examples{
pool_groups(x1 = 10, x2 = 12, sd1 = 4, sd2 = 5, n1 = 10, n2 = 11)

}
