% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{r_to_sdd}
\alias{r_to_sdd}
\title{Get change score standard deviation from pre-post correlation}
\usage{
r_to_sdd(sd_pre, sd_post, r)
}
\arguments{
\item{sd_pre, sd_post}{pre and post standard deviations.}

\item{r}{pre-post correlation.}
}
\value{
A numeric, the standard deviation of change scores.
}
\description{
Calculates the change score standard deviation of a group from the pre and
post standard deviations and the pre-post correlation.
}
\examples{
r_to_sdd(sd_pre = 5, sd_post = 7, r = 0.9)

}
