% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sd_avg}
\alias{sd_avg}
\title{Calculate average standard deviation}
\usage{
sd_avg(sd1, sd2)
}
\arguments{
\item{sd1, sd2}{standard deviations of the first and second group.}
}
\value{
the average standard deviation
}
\description{
Averages the standard deviations of two groups (square root of average
variance).
}
\examples{
sd_avg(5, 3)

}
