% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Calculate pooled standard deviation}
\usage{
sd_pooled(sd1, sd2, n1, n2, mle = FALSE)
}
\arguments{
\item{sd1, sd2}{standard deviations of the first and second group.}

\item{n1, n2}{sample sizes of the first and second group.}

\item{mle}{Whether the maximum likelihood estimator should be used, which has
N instead of df (N-2) in the denominator. See for example McGrath & Mayer
(2006). Defaults to FALSE.}
}
\value{
the pooled standard deviation
}
\description{
Pools the standard deviation of two groups
}
\examples{
sd_pooled(5, 3, 15, 20)

}
