% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sdd_to_r}
\alias{sdd_to_r}
\title{Get pre-post correlation from change score standard deviation}
\usage{
sdd_to_r(sd_pre, sd_post, sd_d)
}
\arguments{
\item{sd_pre, sd_post}{pre and post standard deviations.}

\item{sd_d}{standard deviation of change scores}
}
\value{
A numeric, the pre-post correlation.
}
\description{
Calculates the pre-post correlation for a group from the pre, post and change
score standard deviation.
}
\examples{
sdd_to_r(sd_pre = 3, sd_post = 4, sd_d = 2)

}
