% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es.R
\name{smcr}
\alias{smcr}
\title{Calculating a standardized change score (raw score based standardization)}
\usage{
smcr(xpre, xpost, sdpre, sdpost, n, r, hedges = TRUE, exact = TRUE)
}
\arguments{
\item{xpre, xpost}{mean of the pre and post score}

\item{sdpre, sdpost}{standard deviation of the pre and post score}

\item{n}{sample size}

\item{r}{pre-post correlation}

\item{hedges}{whether an bias correction for small sample sizes according to
Hedges should be applied (defaults to TRUE).}

\item{exact}{whether to use the exact formula for the small sample size
correction (otherwise an approximation is used). This is only relevant when
\code{hedges=TRUE}. Defaults to TRUE.}
}
\description{
Calculates the standardized change score with raw score standardization for a
single group.
}
\examples{
smcr(10, 12, 2, 3, 20, 0.8)

}
